Worst case scenario is the string contains size_t characters that are all the same

+++ Code/Firecode/FirstNonDuplicateCharacter.cpp
@@ -15,7 +15,7 @@
 
 char firstNonRepeating(std::string str)
 {
-  std::unordered_map<char, unsigned int> hm;
+  std::unordered_map<char, size_t> hm;
 
   // Count occurrences of each character
   for (const auto c : str)
